+++ src/Components/Navbar/Navbar.jsx
@@ -13,8 +13,7 @@ import Tooltip from "@mui/material/Tooltip";
 import MenuItem from "@mui/material/MenuItem";
 import PetsIcon from "@mui/icons-material/Pets";
 import { useMediaQuery } from "@mui/material";
-import { loadStripe } from '@stripe/stripe-js';
-import axios from 'axios';
+
 // Jotai
 import { useSetAtom, useAtomValue } from "jotai";
 import { currentUserAtom } from "../../Atoms/currentuser";
@@ -57,33 +56,7 @@ function Navbar() {
         setUserId(null);
     };
 
-    const handlePayment = async () => {
-      try {
-        // Make the API request to create a checkout session
-        const response = await fetch('http://localhost:3000/checkout', {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({ amount: 100 }), // Adjust the amount according to your needs
-        });
     
-        // Handle the API response
-        if (response.ok) {
-          const { sessionId } = await response.json();
-    
-          // Load the Stripe instance
-          const stripe = await loadStripe('pk_test_51MeC2IGFwxWB3B7MgBO0JGpTwYeudavjvvMa9Jbcu4PN1gerYAI9ErBoETZPAKKbF21EHxu8MUfATXaTvSrZgYSj00YrK3BnKT'); 
-    
-          // Redirect the user to the Stripe checkout page
-          stripe.redirectToCheckout({ sessionId });
-        } else {
-          console.error('Error:', response.status);
-        }
-      } catch (error) {
-        console.error('Error:', error);
-      }
-    };
 
     const isMobileScreen = useMediaQuery("(max-width: 960px)");
